feat(draft): create scarefold for draft
- add route, middleware and controller
