+++ src/routes/index.js
@@ -6,6 +6,7 @@ import twitterRouter from './twitter';
 import projectRouter from './project';
 import fileRouter from './files';
 import imageRouter from './images';
+import draftRouter from './drafts';
 
 const router = Router();
 
@@ -15,5 +16,6 @@ router.use('/twitter', twitterRouter);
 router.use('/project', projectRouter);
 router.use('/files', fileRouter);
 router.use('/images', imageRouter);
+router.use('/drafts', draftRouter);
 
 export default router;
